Add pim search to quick access (#209)
* feat(quick-access): support PIM Search WIP

* feat(quick-access): only search products on projects with activated pim-search

* feat(test-utils): enable mocking sdk responses

So far it was quite hard / not possible to mock responses of sdk actions processed by our sdk middleware. This commit enables support for sdkMocks in the options of renderWithRedux.

* feat(quick-access): add support for pim search

This commit enables support for PIM search in quick-access. It is now possible to search for products using the fullText search. This is only available in projects indexed by pim-indexer.

* fix(quick-access): use separate id for pim-search results

* test(quick-access): remove unnecessary mock factories

* feat(test-utils): allow faking failing sdk request

* fix(quick-access): pass debounce args to debounce fn

* docs(quick-access): add isProjectPimIndexed explanation

* refactor(quick-access): replace boolean with dedicated pimIndexerStates