refactor(app-shell): to not `ReconfigureFlopflip` but pass `projectKey` from url (#314)
#### Summary

This pull request refactors the app-shell and the setup-flopflip components to not reconfigure flopflip but to pass the projectKey upon first render.

#### Description

I received feedback about flashings of feature toggles in some parts of the app. Furthermore, in certain navigation scenarios an upcoming feature to not show up.

After spending a reasonable amount of time again to understand how our app renders I suggest to simplify the setup:

- Use the `selectProjectKeyFromUrl` which is used in app-shell
- Pass the projectKey from the app-shell to the setup-flopflip component
- To not `ReconfigureFlopflip` at all

Another option would have been to remove the memoization on `ReconfigureFlopflip` to make sure there is no reconfiguration which is "invalidated" right after due to the the app-shell re-rendering `ConfigureFlopflip`.

I hope this simplification makes sense.