feat(app-shell): add allowing to pass ldTrackingTenant alone (#528)
#### Summary

This pull request adds the ability to add the `launchdarklyTrackingTenant` alone without the other user information.

Today we learned that sign up will not be available on all environments at the same time. Given this needs to be toggled, we need to be able to at least specify the `launchdarklyTrackingTenant`. When running the sign up application we don't have any other user information (as there is no user). 
Before we assumed that we can "get way" with toggling for all environments we deploy to equally.